- Alguns ajustes na implementação de IUnitOfWork
- Modificações no cadastro de lojas para novas tentativas com o MVC. Algumas coisas podem ter parado de funcionar

+++ src/Graxei/Graxei.Aplicacao.Implementacao.MVC4Unity/Areas/Administrativo/Models/EnderecoIndiceModel.cs
@@ -5,7 +5,7 @@ namespace Graxei.Aplicacao.Implementacao.MVC4Unity.Areas.Administrativo.Models
     public class EnderecoIndiceModel
     {
         public int IdLista { get; set; }
-        public long IdEstado { get; set; }
+        public int IdEstado { get; set; }
         public Endereco Endereco { get; set; }
     }
 }
